chaos tests to nightly
