+++ solr/core/src/test/org/apache/solr/cloud/ChaosMonkeySafeLeaderTest.java
@@ -16,18 +16,12 @@
  */
 package org.apache.solr.cloud;
 
-import org.apache.lucene.tests.util.LuceneTestCase.Slow;
-
 /*
  * Implementation moved to AbstractChaosMonkeySafeLeaderTestBase as is used by HDFS contrib module test
  */
-@Slow
 public class ChaosMonkeySafeLeaderTest extends AbstractChaosMonkeySafeLeaderTestBase {
-
-  private static final String DIRECTORY_FACTORY = "solr.StandardDirectoryFactory";
-
   @Override
   protected String getDirectoryFactory() {
-    return DIRECTORY_FACTORY;
+    return "solr.StandardDirectoryFactory";
   }
 }
